pyproject.toml
[project]
name = "aws-local-dev"
version = "0.1.0"
description = "Local AWS Development Environment"
requires-python = ">=3.12.3"
dependencies = [
    "localstack>=3.2.0",
    "boto3>=1.34.69",
    "aws-cdk-lib>=2.131.0",
    "python-dotenv>=1.0.1",
    "typer>=0.11.0",
    "rich>=13.7.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.1.1",
    "ruff>=0.3.3",
    "mypy>=1.9.0",
    "boto3-stubs[essential]>=1.34.69",
    "pre-commit>=3.6.2",
]

[build-system]
requires = ["setuptools>=68.0.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]
include = ["aws_local_dev*"]

[tool.setuptools.package-data]
aws_local_dev = ["py.typed"]

[tool.ruff]
target-version = "py312"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "B", "C4", "SIM", "TCH", "UP", "N", "S", "A"]

[tool.ruff.lint.per-file-ignores]
"tests/**/*" = ["S101"]  # Ignore assert statements in test files

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
line-ending = "auto"

[tool.mypy]
python_version = "3.12"
strict = true
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
explicit_package_bases = true
namespace_packages = true